pyproject.toml
[tool.poetry]
name = "gmg"
version = "0.3.2"
description = "Public forum scrapper"
authors = ["Alexandre D. Díaz <dev@redneboa.es>"]
include = [
    "data/*.xml",
    "gmgl/**/*.py",
    "migrations/",
    "routes/*.py",
    "static/",
    "templates/**/*.j2",
    "tools/*.py",
    "translations/**/*.mo",
    "translations/**/*.po",
    "babel.cfg",
    "cli.py",
    "gmg.py",
    "gunicorn.conf.py",
    "package-lock.json",
    "package.json",
    "poetry.lock",
    "postcss.config.js",
    "pypoject.toml",
    "rollup.config.js",
    "run.py",
    "tailwind.config.js",
    "uwsgi.ini",
]

[tool.poetry.dependencies]
python = "^3.10"
cssselect = "^1.1.0"
distro = "^1.6.0"
Flask = "^2.0.2"
Flask-APScheduler = "^1.12.2"
Flask-Assets = "^2.0"
Flask-Babel = "^2.0.0"
Flask-Compress = "^1.11"
Flask-Cors = "^3.0.10"
Flask-SQLAlchemy = "^2.5.1"
Flask-WTF = "^1.0.0"
Flask-Migrate = "^3.1.0"
flask-csp = "^0.10"
gevent = "^21.12.0"
gunicorn = "^20.1.0"
html2text = "^2020.1.16"
webassets-rollup = "^1.0.0"
PyYAML = "^5.4.1"
requests = "^2.27.1"
Pillow = "^8.4.0"
lxml = "^4.7.1"
psycopg2-binary = "^2.9.3"
python-dotenv = "^0.20.0"
jinja2 = "3.0.3"
spacy = "^3.3.0"
pytest-mock = "^3.7.0"
Flask-SQLAlchemy-Caching = "^1.0.4"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
coverage = "^6.3.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
